requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
bzip2=1.0.8=h6c40b1e_6
ca-certificates=2024.9.24=hecd8cb5_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
filelock=3.16.1=pypi_0
fsspec=2024.10.0=pypi_0
huggingface-hub=0.26.2=pypi_0
idna=3.10=pypi_0
jinja2=3.1.4=pypi_0
libffi=3.4.4=hecd8cb5_1
markupsafe=3.0.2=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=hcec6c5f_0
networkx=3.4.2=pypi_0
numpy=1.26.4=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.0.15=h46256e1_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pillow=11.0.0=pypi_0
pip=24.2=py310hecd8cb5_0
python=3.10.15=hce00570_1
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=hca72f7f_0
requests=2.32.3=pypi_0
safetensors=0.4.5=pypi_0
scipy=1.14.1=pypi_0
setuptools=75.1.0=py310hecd8cb5_0
six=1.16.0=pypi_0
sqlite=3.45.3=h6c40b1e_0
sympy=1.13.3=pypi_0
timm=1.0.11=pypi_0
tk=8.6.14=h4d00af3_0
torch=2.2.2=pypi_0
torchaudio=2.2.2=pypi_0
torchvision=0.17.2=pypi_0
tqdm=4.67.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
wheel=0.44.0=py310hecd8cb5_0
xz=5.4.6=h6c40b1e_1
zlib=1.2.13=h4b97444_1
